Make highlight -> make-bulletin-separator obsolescence clearer in the code
I'm still not 100% clear on how this ought to be done.
Making this change because I was wondering if I'd skipped the obsolescence
commit somehow (ie I couldn't tell).

It seems like you need to call make-obsolete after both functions have been
defined, and one of the functions of make-obsolete is to add a note to the
docstring of the old function. So I haven't added that note myself, although it
would be very helpful to a reader of the code.

Ideally I'd like

(defun make-bulletin-separator ...)
(make-obsolete 'highlight-80-dashes 'make-bulletin-separator)
(defun highlight-80-dashes ...)

But the highlight symbol isn't yet defined. (Maybe I should defer its evaluation
somehow?) Anyway, this is kind of nitpicky, especially given that I can delete
the 80-dash functions instead of obsoleting them. But I don't wanna lose that
code just yet~ and if it's gone but preserved in Git then I'll never remember to
forget to it.

Every LoC is sacred, when you're starting out.