Rename bulletin[s]-mode, minor->major mode, etc
Also add filetype matching and customizable font (or the start thereof).

I've hit that point of the night where too many things have happened to break
down cleanly into commits. Ceeeeebs.

But it *works*, so that's cool. Current status is that it feels like a very
under-featured package, but what it does do, it does well. Feels good.

Filetype match is from:
https://www.emacswiki.org/emacs/AutoModeAlist